Ajout des tests unitaires pour le backend avec Jest et Supertest
